refactor: by making JwtLogoutFilter which extends LogoutFilter
