+++ src/main/java/com/mapu/global/common/config/SecurityConfig.java
@@ -1,13 +1,8 @@
 package com.mapu.global.common.config;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mapu.global.common.response.BaseResponse;
-import com.mapu.global.jwt.JwtFilter;
-import com.mapu.global.jwt.JwtLogoutHandler;
+import com.mapu.global.jwt.api.JwtFilter;
+import com.mapu.global.jwt.api.JwtLogoutFilter;
 import com.mapu.global.jwt.application.JwtService;
-import com.mapu.global.jwt.JwtUtil;
-import com.mapu.global.jwt.dao.JwtRedisRepository;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -16,7 +11,7 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.security.web.authentication.logout.LogoutFilter;
 
 
 @Configuration
@@ -24,9 +19,7 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-    private final JwtUtil jwtUtil;
     private final JwtService jwtService;
-    private final JwtRedisRepository jwtRedisRepository;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -39,28 +32,14 @@ public class SecurityConfig {
         http
                 .formLogin((auth) -> auth.disable());
 
-        http
-                .logout((logout) ->
-                        logout.logoutRequestMatcher(new AntPathRequestMatcher("/logout", "POST"))
-                                .deleteCookies(jwtUtil.REFRESH)
-                                .addLogoutHandler(new JwtLogoutHandler(jwtService))
-                                .logoutSuccessHandler((request, response, authentication) -> {
-                                    response.setStatus(HttpServletResponse.SC_OK);
-                                    response.setContentType("application/json;charset=UTF-8");
-                                    BaseResponse<String> baseResponse = new BaseResponse<>("로그아웃 성공");
-                                    ObjectMapper objectMapper = new ObjectMapper();
-                                    String jsonResponse = objectMapper.writeValueAsString(baseResponse);
-                                    response.getWriter().write(jsonResponse);
-                                }));
-
         //HTTP Basic 인증 방식 disable
         http
                 .httpBasic((auth) -> auth.disable());
 
         //JWTFilter 추가
         http
-                .addFilterBefore(new JwtFilter(jwtService), UsernamePasswordAuthenticationFilter.class);
-                //.addFilterBefore(new CustomLogoutFilter(jwtUtil, jwtRedisRepository), LogoutFilter.class);
+                .addFilterBefore(new JwtFilter(jwtService), UsernamePasswordAuthenticationFilter.class)
+                .addFilterAt(new JwtLogoutFilter(jwtService), LogoutFilter.class);
 
         //oauth2 -> 필요없음
 //        http
