+++ src/main/java/com/mapu/global/jwt/application/JwtLogoutHandler.java
@@ -1,8 +1,8 @@
-package com.mapu.global.jwt;
+package com.mapu.global.jwt.application;
 
 import com.mapu.global.common.exception.BaseException;
 import com.mapu.global.common.exception.errorcode.BaseExceptionErrorCode;
-import com.mapu.global.jwt.application.JwtService;
+import com.mapu.global.jwt.JwtUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
