+++ src/main/java/com/mapu/global/jwt/application/JwtLogoutSuccessHandler.java
@@ -1,0 +1,23 @@
+package com.mapu.global.jwt.application;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mapu.global.common.response.BaseResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+
+import java.io.IOException;
+
+public class JwtLogoutSuccessHandler implements LogoutSuccessHandler {
+    @Override
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.setContentType("application/json;charset=UTF-8");
+        BaseResponse<String> baseResponse = new BaseResponse<>("로그아웃 성공");
+        ObjectMapper objectMapper = new ObjectMapper();
+        String jsonResponse = objectMapper.writeValueAsString(baseResponse);
+        response.getWriter().write(jsonResponse);
+    }
+}
